chore(docs): cleans up docs a bit more for final hub site
