+++ packages/client/src/index.ts
@@ -69,9 +69,9 @@ export class Client {
    * Create a new gRPC client instance from a supplied key and secret
    * @param key The KeyInfo object containing {key: string, secret: string, type: 0}. 0 === User Group Key, 1 === Account Key
    */
-  static async withUserKey(key: KeyInfo, host = defaultHost, debug = false) {
+  static async withKeyInfo(key: KeyInfo, host = defaultHost, debug = false) {
     const context = new Context(host, debug)
-    await context.withUserKey(key)
+    await context.withKeyInfo(key)
     return new Client(context)
   }
