+++ packages/security/src/index.ts
@@ -10,10 +10,12 @@ import multibase from 'multibase'
 
 /**
  * UserAuth is a type describing the minimal requirements create a session from a user group key. Generate with {@link createUserAuth}.
+ *
+ * Optional token generated by {@link https://textileio.github.io/js-threads/classes/_textile_threads_client.client.html#gettoken | Client.getToken} or {@link https://textileio.github.io/js-threads/classes/_textile_threads_client.client.html#gettoken | Client.getTokenChallenge}.
  * @public
  * @example
  * Import
- * ```
+ * ```typescript
  * import {UserAuth} from '@textile/hub';
  * ```
  * @param {string} key - API key. Can be embedded/shared within an app.
@@ -29,11 +31,11 @@ export type UserAuth = {
 }
 
 /**
- * KeyInfo is a type that contains the API Secret. It should never be shared in insecure environments.
+ * KeyInfo is a type that contains the API Secret. It should never be shared in insecure environments. Type 0=ACCOUNT, 1=USER.
  * @public
  * @example
  * Import
- * ```
+ * ```typescript
  * import {KeyInfo} from '@textile/hub';
  * ```
  * @param {string} key - API key. Can be embedded/shared within an app.
@@ -58,6 +60,11 @@ export type KeyInfo = {
 /**
  * Create an API signature for use in authenticated systems. Generate with {@link createAPISig}.
  * @public
+ * @example
+ * Import
+ * ```typescript
+ * import {APISig} from '@textile/hub';
+ * ```
  * @param {string} sig - The signature of the authentication message.
  * @param {string} msg - The authentication message.
  */
@@ -72,18 +79,20 @@ export type APISig = {
  * This function should NOT be used client-side, as it requires a key secret.
  * @public
  * @example
- * Import
- * ```
- * import {createAPISig} from '@textile/hub';
+ * Basic usage
+ * ```typescript
+ * import {createAPISig, APISig} from '@textile/hub'
+ *
+ * const sig: APISig = createAPISig(USER_API_KEY)
  * ```
  * @param {string} secret - The key secret to generate the signature. See KeyInfo for details.
  * @param {Date} date - An optional future Date to use as signature message. Once `date` has passed, this
  * authorization signature and message will expire. Defaults to one minute from `Date.now`.
  */
-export const createAPISig = async (
+export async function createAPISig(
   secret: string,
   date: Date = new Date(Date.now() + 1000 * 60),
-): Promise<APISig> => {
+): Promise<APISig> {
   const sec = multibase.decode(secret)
   const msg = (date ?? new Date()).toISOString()
   const hash = new HMAC(sec)
@@ -101,7 +110,7 @@ export const createAPISig = async (
  * @public
  * @example
  * Create a new UserAuth
- * ```
+ * ```typescript
  * import {createAPISig, Client, KeyInfo, UserAuth} from '@textile/hub';
  *
  * // The first step is to create a basic session with your user group keys. See KeyInfo type.
@@ -110,7 +119,7 @@ export const createAPISig = async (
  *   secret: USER_API_SECRET,
  *   type: 0,
  * }
- * const db = await Client.withUserKey(keyInfo)
+ * const db = await Client.withKeyInfo(keyInfo)
  *
  * // Create an expiration and create a signature. 60s or less is recommended.
  * const expiration = new Date(Date.now() + 60 * seconds)
@@ -122,12 +131,12 @@ export const createAPISig = async (
  * @param {Date} date - An optional future Date to use as signature message. Default 1 minute from now.
  * @param {string} token - An optional user API token.
  */
-export const createUserAuth = async (
+export async function createUserAuth(
   key: string,
   secret: string,
   date: Date = new Date(Date.now() + 1000 * 60),
   token?: string,
-): Promise<UserAuth> => {
+): Promise<UserAuth> {
   const partial = await createAPISig(secret, date)
   return {
     ...partial,
@@ -141,5 +150,5 @@ export const createUserAuth = async (
  * @public
  */
 export const expirationError = new Error(
-  'Auth expired. Consider calling withUserKey or withAPISig to refresh.',
+  'Auth expired. Consider calling withKeyInfo or withAPISig to refresh.',
 )
